"1:30pm.
```
#### Exercise `⇔≃×` (recommended)

Show that `A ⇔ B` as defined [earlier][plfa.Isomorphism#iff]
is isomorphic to `(A → B) × (B → A)`
```

Let me deal with this exercise.

All the exercises in PLFA have just been so easy so far. This is the complete opposite of SF. Here exercises are in fact too easy.

1:45pm.

```
⇔≃× : ∀ (A B) → A ⇔ B ≃ (A → B) × (B → A)
⇔≃× A B =
  record
    { from = λ ⟨ a , b ⟩ →
        record
          { from = b
          ; to = a
          }
    ; to = {!!}
    ; from∘to = {!!}
    ; to∘from = {!!}
    }
```
```
E:\TheoremProving\plfa\Connectives.lagda:246,18-250,12
(a : _42) (, : _46 (A = A) (B = B) (⟨ = ⟨) (a = a))
(b : _50 (A = A) (B = B) (⟨ = ⟨) (a = a) (, = ,))
(⟩ : _54 (A = A) (B = B) (⟨ = ⟨) (a = a) (, = ,) (b = b)) →
_56 (A = A) (B = B) (⟨ = ⟨) (a = a) (, = ,) (b = b) (⟩ = ⟩)
!=< A ⇔ B of type
piSort _41
(λ a →
   piSort (_45 (A = A) (B = B) (⟨ = ⟨) (a = a))
   (λ , →
      piSort (_49 (A = A) (B = B) (⟨ = ⟨) (a = a) (, = ,))
      (λ b →
         piSort (_53 (A = A) (B = B) (⟨ = ⟨) (a = a) (, = ,) (b = b))
         (λ _ → Set))))
when checking that the expression
λ a , b ⟩ → record { from = b ; to = a } has type A ⇔ B
```

Ugh, these type errors...

1:55pm.

```
open import plfa.Isomorphism using (_⇔_)
open _⇔_

⇔≃× : ∀ (A B) → A ⇔ B ≃ (A → B) × (B → A)
⇔≃× A B =
  record
    { from = λ {⟨ x , y ⟩ →
        record
          { from = y
          ; to = x
          }
        }
    ; to = λ A⇔B → ⟨ to A⇔B , from A⇔B ⟩
    ; from∘to = λ x → refl
    ; to∘from = λ {⟨ x , y ⟩ → refl }
    }
```

Done with this. It seems I have some work to do outside so let me take a break from this here.

I'll get back to this later.

This particular chapter seems like it will have plenty of exercises. I won't be able to deal with it all today. But at least I am more than halfway through the book."