"Time for union commutativity.
```
⊎-comm : ∀ A B →  A ⊎ B ≃ B ⊎ A
```

Here is the definition. Now how do I construct thme again?

1:10pm.

```
    { to = λ {
         inj₁ x y → ?
         inj₂ x y → ?
         }
```

Damn, what is the syntax for this in lambdas?

1:15pm.

```
⊎-comm : ∀ {A B} →  A ⊎ B ≃ B ⊎ A
⊎-comm =
  record
    { to = λ { (inj₁ x) → inj₂ x
             ; (inj₂ x) → inj₁ x
             }
    ; from = λ { (inj₁ x) → inj₂ x
               ; (inj₂ x) → inj₁ x
               }
    ; to∘from = λ { (inj₁ x) → refl
                  ; (inj₂ x) → refl
                  }
    ; from∘to = λ { (inj₁ x) → refl
                  ; (inj₂ x) → refl
                  }
    }
```

The next is associativity.

```
⊎-assoc : ∀ {A B C} →  A ⊎ (B ⊎ C)  ≃ (A ⊎ B) ⊎ C
⊎-assoc =
  record
    { to = λ { (inj₁ x) → inj₁ (inj₁ x)
             ; (inj₂ (inj₁ x)) → inj₁ (inj₂ x)
             ; (inj₂ (inj₂ x)) → inj₂ x
             }
    ; from = λ { (inj₁ (inj₁ x)) → inj₁ x
               ; (inj₁ (inj₂ x)) → inj₂ (inj₁ x)
               ; (inj₂ x) → inj₂ (inj₂ x)
               }
    ; to∘from = λ { (inj₁ (inj₁ x)) → refl
                  ; (inj₁ (inj₂ x)) → refl
                  ; (inj₂ x) → refl
                  }
    ; from∘to = λ { (inj₁ x) → refl
                  ; (inj₂ (inj₁ x)) → refl
                  ; (inj₂ (inj₂ x)) → refl
                  }
    }
```

1:40pm. Done. What is next?

```
#### Exercise `⊥-identityˡ` (recommended)

Show empty is the left identity of sums up to isomorphism.

\begin{code}
-- Your code goes here
\end{code}

#### Exercise `⊥-identityʳ`

Show empty is the right identity of sums up to isomorphism.

\begin{code}
-- Your code goes here
\end{code}
```

Compared to what I have to deal with in Coq, this stuff is so trivial.

```
⊤-identityˡ : ∀ {A : Set} → ⊤ × A ≃ A
⊤-identityʳ : ∀ {A : Set} → (A × ⊤) ≃ A
```

1:50pm.

```
⊥-identityˡ : ∀ {A : Set} → ⊥ × A ≃ A
⊥-identityʳ : ∀ {A : Set} → (A × ⊥) ≃ A
```

Let me prove these two.

...Let me read the new Baki chapter first though.

1:55pm. That was short, let me do this thing. I have to admit it is really distracting to have to write out all these symbols, but that is something I will just have to get used to.

2:05pm.

```
⊥-identityˡ : ∀ {A : Set} → ⊥ ⊎ A ≃ A
⊥-identityˡ =
  record
    { to = λ { (inj₁ ())
             ; (inj₂ x) → x
             }
    ; from = inj₂
    ; to∘from = λ x → refl
    ; from∘to = λ { (inj₁ ())
                  ; (inj₂ x) → refl
                  }
    }

⊥-identityʳ : ∀ {A : Set} → (A ⊎ ⊥) ≃ A
⊥-identityʳ =
  record
    { to = λ { (inj₂ ())
             ; (inj₁ x) → x
             }
    ; from = inj₁
    ; to∘from = λ x → refl
    ; from∘to = λ { (inj₂ ())
                  ; (inj₁ x) → refl
                  }
    }
```

Here it is.

I am 2/3rds done with the chapter. Let me continue moving forward.

2:20pm. Got a PR by Anti. I thought it would be small, but he did quite a bit it seems.

...Since I am 75% done with the chapter, let me finish it and then I'll go through it.

`keywords's argument names.` should be `keywords' argument names.`

Let me just note this for later. Ah, I can put it as a note directly in the PR. It is my first time doing a PR myself so I have a bit to learn.

2:35pm.

```
currying : ∀ {A B C : Set} → (A → B → C) ≃ (A × B → C)
currying =
  record
    { to      =  λ{ f → λ{ ⟨ x , y ⟩ → f x y }}
    ; from    =  λ{ g → λ{ x → λ{ y → g ⟨ x , y ⟩ }}}
    ; from∘to =  λ{ f → refl }
    ; to∘from = λ{ g → extensionality λ{ ⟨ x , y ⟩ → refl }}
    }
```

I had to study this last case for a bit. It is not obvious at first glance what extensionality is supposed to be doing in this case, but I think I get it.

Let me move on.

This would be easy in Coq as all the steps would be explicit, but here I needed to think about it for a bit.

```
Corresponding to the law

    p ^ (n + m) = (p ^ n) * (p ^ m)

we have the isomorphism:

    (A ⊎ B) → C  ≃  (A → C) × (B → C)
```

I admit this is mindblowing. I had no idea that implication has the same rules as exponentiation in regular arithmetic. I have no idea what to make of it which is why I haven't commented on it before.

2:40pm. Er, I need to take off for a bit here. I do not think the housework should be as time consuming as it was yesterday."