"3:10pm. Let me resume. I am still not done with chores, but that one thing still left I need to do later.
3:20pm.

```
×-distrib-⊎ : ∀ {A B C : Set} → (A ⊎ B) × C ≃ (A × C) ⊎ (B × C)
```

This really is mindblowing. If implication corresponds to exponentiation, I wonder what would correspond to the logarithm? There might some links with Bayesian inference here.

3:30pm.

```
#### Exercise `⊎-weak-×` (recommended)

Show that the following property holds:
\begin{code}
postulate
  ⊎-weak-× : ∀ {A B C : Set} → (A ⊎ B) × C → A ⊎ (B × C)
\end{code}
This is called a _weak distributive law_. Give the corresponding
distributive law, and explain how it relates to the weak version.

\begin{code}
-- Your code goes here
\end{code}
```

I had to think about this for a bit as it is giving sparks. I really am starting to understand why in that pumping lemma case I did not know how to do the split. This is really giving me an insight into what implication really is. I really am too green. There is really no end to the depth of functional programming.

Let me do this exercise.

3:45pm.

```
This is called a _weak distributive law_. Give the corresponding
distributive law, and explain how it relates to the weak version.

\begin{code}
weak-distr-embeds-strong : ∀ {A B C : Set} → (A ⊎ B) × C ≲ (A × C) ⊎ (B × C)
\end{code}
```

Let me make the explanation a proof.

3:55pm.

```
weak-distr-embeds-strong : ∀ {A B C : Set} → (A ⊎ B) × C ≲ (A × C) ⊎ (B × C)
weak-distr-embeds-strong =
  record
    { to     = λ { ⟨ inj₁ x , b ⟩ → inj₁ ⟨ x , b ⟩ ; ⟨ inj₂ x , b ⟩ → inj₂ ⟨ x , b ⟩}
    ; from   = λ { (inj₁ ⟨ x , x₁ ⟩) → ⟨ inj₁ x , x₁ ⟩ ; (inj₂ ⟨ x , x₁ ⟩) → ⟨ inj₂ x , x₁ ⟩}
    ; from∘to = λ { ⟨ inj₁ x , x₁ ⟩ → refl ; ⟨ inj₂ x , x₁ ⟩ → refl}
    }
```

Let me try to prove an isomorphism as an exercise here.

I do not think it should pass.

4pm.

```
weak-distr-embeds-strong : ∀ {A B C : Set} → (A ⊎ B) × C ≲ (A × C) ⊎ (B × C)
weak-distr-embeds-strong =
  record
    { to     = λ { ⟨ inj₁ x , b ⟩ → inj₁ ⟨ x , b ⟩ ; ⟨ inj₂ x , b ⟩ → inj₂ ⟨ x , b ⟩}
    ; from   = λ { (inj₁ ⟨ x , x₁ ⟩) → ⟨ inj₁ x , x₁ ⟩ ; (inj₂ ⟨ x , x₁ ⟩) → ⟨ inj₂ x , x₁ ⟩}
    ; from∘to = λ { ⟨ inj₁ x , x₁ ⟩ → refl ; ⟨ inj₂ x , x₁ ⟩ → refl}
    }

weak-distr-≃-strong : ∀ {A B C : Set} → (A ⊎ B) × C ≃ (A × C) ⊎ (B × C)
weak-distr-≃-strong =
  record
    { to     = λ { ⟨ inj₁ x , b ⟩ → inj₁ ⟨ x , b ⟩ ; ⟨ inj₂ x , b ⟩ → inj₂ ⟨ x , b ⟩}
    ; from   = λ { (inj₁ ⟨ x , x₁ ⟩) → ⟨ inj₁ x , x₁ ⟩ ; (inj₂ ⟨ x , x₁ ⟩) → ⟨ inj₂ x , x₁ ⟩}
    ; from∘to = λ { ⟨ inj₁ x , x₁ ⟩ → refl ; ⟨ inj₂ x , x₁ ⟩ → refl}
    ; to∘from = λ { (inj₁ ⟨ x , x₁ ⟩) → refl ; (inj₂ ⟨ x , x₁ ⟩) → refl}
    }
```

Whops, I am proving the wrong thing here.

Let me do it all again.

```
weak-distr-embeds-strong : ∀ {A B C : Set} → A ⊎ (B × C) ≲ (A × C) ⊎ (B × C)
weak-distr-embeds-strong =
  record
    { to     = {!!}
    ; from   = {!!}
    ; from∘to = {!!}
    }
```

This is what I should be proving.

4pm.

```
weak-distr-embeds-strong : ∀ {A B C : Set} → A ⊎ (B × C) ≲ (A × C) ⊎ (B × C)
weak-distr-embeds-strong =
  record
    { to     = λ { (inj₁ x) → {!!} ; (inj₂ x) → {!!}}
    ; from   = {!!}
    ; from∘to = {!!}
    }
```
```
Goal: A × C ⊎ B × C
————————————————————————————————————————————————————————————
x : A
```

I thought it would be an embedding, but it is not even that.

```
-- As `(A × C) ⊎ (B × C)` is isomoprhic to `(A ⊎ B) × C`, it  implies `A ⊎ (B × C)`,
-- but the converse does not hold.
```

I'll give this as the answer.

I am not sure if it possible to prove that the converse does not hold...

...No I do not see how.

4:10pm. Let me read Omega and then I'll do the last exercise. I already did it by accident, but there is also the question that I need to think about.

```
Does the converse hold? If so, prove; if not, give a counterexample.

\begin{code}
-- Your code goes here
\end{code}
```

Let me think.

`(A ⊎ C) × (B ⊎ D) → (A × B) ⊎ (C × D)`

Yeah, what if I get `C x B` on the left side? Then the implication would not hold.

4:20pm. Finally done with the chapter. Let me take a look at the next one and then I will get to work on that PR."