"5pm. Merged the PR. I said I would look into the next chapter, so let me do that now.
5:05pm. https://plfa.github.io/Fonts/

Oh, I've been wondering about the font spacing. I need to get a font that is monospaced even for special characters.

https://plfa.github.io/Negation/

Including this one only 4 chapters are left. It has been a smooth ride so far.

I am going to skip part two for the time being. I am not interested in PL foundations that much at the moment - I just want to beat the Coq book.

After I do these 4 chapters, I will have the basics of Agda under my belt and will be able to study that proof by gallais. Once I understand it, I will translate it to Coq and will be able to move on.

5:35pm. I skimmed the chapter. One of the last exercises is the 5 star logic exercise from the Coq SF book. That will be a challenge, but it already one I beat, so I should have no trouble porting it to Agda. Actually, I overdid it back then - I only needed to make a circular path to prove all the equivalences, so I will make it concise this time.

5:40pm. https://github.com/gallais/potpourri/blob/master/agda/poc/PigeonHole.agda

It does seem like I understand all the syntax now, but I'll leave studying it for a few days later.

Once the pigeonhole proof is out of the way, that will only leave that 4 star problem I am currently stuck on. I might have to beg for a hint to that one as well, but I will give it a good honest try first. I'll try the other cases as well before that.

It really is likely that the IndProp chapter might have singular difficulty out of all of them.

My goal is to beat Logical Foundations. I'll consider whether I want to do PL Foundations after that."