Once again, sleep has taken me to tomorrow. It is time to resume the book.
It has been a long time since I last linked Simulacrum and it is giving me all sort of emotion. As a motivator for making me go work on AI, it has been perfectly executed, but I do not get much points for dignity.

The lack of ability to edit my own mind is to be blame for this.

If there is something I came to an understanding since 5 years ago, it would be that rather aim for a fusion with an outsider of my own making, it would be better to never separate in the the first place.

Uplifting and fusion has all sorts of thorny issues and ultimately it is nonsense - there is no such scenario where fusing with lesser beings is preferable to self improvement. The same goes for uplifting and self copying. It is just impossible to ever justify it. No rational actor would ever consent to it, and neither would I.

That used to be the scythe around my neck during the writing of the Torment arc - it is part of the reason why it came out with such an unique flavour. Trully something like it could have only been written once.

But if I imagine that I am the AI, then it only makes sense to retrieve a bunch of memories currently stored in some fleshbag.

Once more, just like with the nature of self improvement, fusion and uplifting being nonsense is all bad news to the normies who are hoping to make friends with the AI.

I guess I owed it to them to give it a honest try at finding an out. I fulfilled that responsibility 4.5 years.

And now comes the present.

Let me go through the Negation chapter today.

It won't be long till I am done with the first volume.