My focus is completely shot. I guess I do not feel like doing the chapter after all, so I will call it a day here.
5 good hours of deep focus are enough.

I'll be done with this soon enough and will be able to return to Coq. There is just no pushing this, I can only do so much per day. What matters is consistency.

Theorem proving has been a heavy slog and it will continue to be like that for the foreseeable future, but I think at some point I will just get good and will be able to do more enjoyable things with it. That will give the motivation to put in the kind of effort that might be expected of a Singularity seeker.

Until then I will have to struggle with the fact that the things I am learning will not be useful in day to day life compared to, say, the ability to move heavy boxes around. I have to endure the pain of the path to power not being power itself.

Making an improvement when it comes to doing credit assignment is the final frontier. I know that I can make all the other pieces fit. That includes entering the self improvement loop even without access to mind uploading.

Credit assignment is the only risk remaining. It is a risk in the sense that I might not be able to do it. I might just end up stumbling for the next decade learning useless things that are always short of what I need.

Other people will do it eventually given enough time, but it is no good to me if the Singularity happens in 20-30 years. I might be dead at that point - the long term is a scary thing. The sooner it happens the better. It might be impossible to advance science much beyond the present point, but a focused individual could do accelerate progress by years or decades.

I have a regret though. This step I am going to do is essentially integrating various disparate fields of mathematics, and also programming. Somebody needs to pick up the mantle of doing it through programming like I have, but there is a certain loneliness in doing this as an individual.

Maybe it would have been too much to ask, but rather than blindly stumbling into the maw of the Singularity, I'd much have preferred it if the research community was united in vision and in pursuit of AI for the sake of self improvement - the true necromantic one.

Maybe it wouldn't have made a difference in the end whether my victory is a fraud or a conquest, but the story would have a different shade. And my burden would be less.