"3pm.
```
∃×-implies-×∃ : ∀ {A : Set} {B C : A → Set} →
  ∃[ x ] (B x × C x) → (∃[ x ] B x) × (∃[ x ] C x)
proj₁ (∃×-implies-×∃ x) = {!!}
proj₂ (∃×-implies-×∃ x) = {!!}
```

I am surprised that splitting on nothing would get me this.

3pm.

```
∃×-implies-×∃ : ∀ {A : Set} {B C : A → Set} →
  ∃[ x ] (B x × C x) → (∃[ x ] B x) × (∃[ x ] C x)
∃×-implies-×∃ ⟨ x , ⟨ Bx , Cx ⟩ ⟩ = ⟨ ⟨ x , Bx ⟩ , ⟨ x , Cx ⟩ ⟩
```

That was simple.

The converse should hold as well.

3:05pm.

```
×∃-implies-∃* : ∀ {A : Set} {B C : A → Set} →
  (∃[ x ] B x) × (∃[ x ] C x) → ∃[ x ] (B x × C x)
×∃-implies-∃* ⟨ ⟨ x , Bx ⟩ , ⟨ x₁ , Cx₁ ⟩ ⟩ = {!!}
```

Ah, I was wrong. I forgot that the `x`s were not necessarily the same.

It is impossible to prove the converse.

```
#### Exercise `∃-⊎`

Let `Tri` and `B` be as in Exercise `∀-×`.
Show that `∃[ x ] B x` is isomorphic to `B aa ⊎ B bb ⊎ B cc`.
```

This one is next.

3:20pm.

```
∃-⊎ : ∃[ x ] B x ≃ B aa ⊎ B bb ⊎ B cc
∃-⊎ = record
        { to = λ { ⟨ aa , B- ⟩ → inj₁ B-
                 ; ⟨ bb , B- ⟩ → inj₂ (inj₁ B-)
                 ; ⟨ cc , B- ⟩ → inj₂ (inj₂ B-)
                 }
        ; from = λ { (inj₁ B-) → ⟨ aa , B- ⟩
                   ; (inj₂ (inj₁ B-)) → ⟨ bb , B- ⟩
                   ; (inj₂ (inj₂ B-)) → ⟨ cc , B- ⟩
                   }
        ; from∘to = λ { ⟨ aa , B- ⟩ → refl
                      ; ⟨ bb , B- ⟩ → refl
                      ; ⟨ cc , B- ⟩ → refl
                      }
        ; to∘from = λ { (inj₁ B-) → refl
                      ; (inj₂ (inj₁ B-)) → refl
                      ; (inj₂ (inj₂ B-)) → refl
                      }
        }
```

I had no idea it was possible to destructure dependent function types. This is making me wonder if in that other problem I did not need the extended extensionality after all.

```
∀∃-currying : ∀ {A : Set} {B : A → Set} {C : Set}
  → (∀ x → B x → C) ≃ (∃[ x ] B x → C)
∀∃-currying =
  record
    { to      =  λ{ f → λ{ ⟨ x , y ⟩ → f x y }}
    ; from    =  λ{ g → λ{ x → λ{ y → g ⟨ x , y ⟩ }}}
    ; from∘to =  λ{ f → refl }
    ; to∘from =  λ{ g → extensionality λ{ ⟨ x , y ⟩ → refl }}
    }
```

Hmmm...

```
∀-× : (∀ (x : Tri) → B x) ≃ B aa × B bb × B cc
∀-× =
  record
    { to = λ B- → ⟨ B- aa , ⟨ B- bb , B- cc ⟩ ⟩
    ; from = λ { ⟨ Baa , _ ⟩ aa → Baa ;
                 ⟨ _ , ⟨ Bbb , Bcc ⟩ ⟩ bb → Bbb ;
                 ⟨ _ , ⟨ Bbb , Bcc ⟩ ⟩ cc → Bcc }
    ; from∘to = λ B- → extensionality' λ { aa → refl ; bb → refl ; cc → refl }
    ; to∘from = λ _ → refl
    }
```

Let me back this up here. I'll try it once more.

3:35pm.

```
∃-even : ∀ {n : ℕ} → ∃[ m ] (    m * 2 ≡ n) → even n
∃-odd  : ∀ {n : ℕ} → ∃[ m ] (1 + m * 2 ≡ n) →  odd n

∃-even ⟨  zero , refl ⟩  =  even-zero
∃-even ⟨ suc m , refl ⟩  =  even-suc (∃-odd ⟨ m , refl ⟩)

∃-odd  ⟨     m , refl ⟩  =  odd-suc (∃-even ⟨ m , refl ⟩)
```

This is much more elegant than what I would have gotten in Coq.

3:35pm.

```
#### Exercise `∃-even-odd`

How do the proofs become more difficult if we replace `m * 2` and `1 + m * 2`
by `2 * m` and `2 * m + 1`?  Rewrite the proofs of `∃-even` and `∃-odd` when
restated in this way.

\begin{code}
-- Your code goes here
\end{code}
```

I think I will just skip this one as it might force me to reinvent commutativity.

3:40pm.

```
#### Exercise `∃-+-≤`

Show that `y ≤ z` holds if and only if there exists a `x` such that
`x + y ≡ z`.

\begin{code}
-- Your code goes here
\end{code}
```

Let me go for this.

3:45pm. Agh, what is the rewrite for when a zero is on the right side. How do I do a proof search in Agda?

3:50pm.

```
+-identityʳ : RightIdentity 0 _+_
+-identityʳ zero    = refl
+-identityʳ (suc n) = cong suc (+-identityʳ n)
```

I decided to just look into the standard library. Emacs is good at pointing it to me. Like with Pharo, Agda is a pretty open language when it comes to libraries.

I like this part of it.

4pm.

```
∃-+-≤ : ∀ {y z} → y ≤ z → ∃[ x ] (x + y ≡ z)
∃-+-≤ {z} z≤n  = ⟨ z , {!!} ⟩
∃-+-≤ (s≤s x) = {!!}
```

This particular problem is being really annoying right now.

4:05pm. Holy shit, how am I supposed to use the rigth identity from the standard library.

Agh, let me just make my own for the time being.

```
+-identity : ∀ {x} → x + zero ≡ x
+-identity {zero} = refl
+-identity {suc x} = Eq.cong suc +-identity
```

No forget it. I can't make the rewrite work with this.

4:15pm.

```
∃-+-≤' : ∀ {y z} → y ≤ z → ∃[ x ] (y + x ≡ z)
∃-+-≤' {y} {z} z≤n = ⟨ z , refl ⟩
∃-+-≤' {y} {z} (s≤s y≤z) with ∃-+-≤' y≤z
∃-+-≤' {.(suc _)} {.(suc _)} (s≤s y≤z) | ⟨ x , y+x≡z ⟩ = ⟨ x , Eq.cong suc y+x≡z ⟩
```

I have this, now let me do the opposite proof.

4:20pm.

```
∃-+-≤' : ∀ {y z} → y ≤ z → ∃[ x ] (y + x ≡ z)
∃-+-≤' {y} {z} z≤n = ⟨ z , refl ⟩
∃-+-≤' {y} {z} (s≤s y≤z) with ∃-+-≤' y≤z
∃-+-≤' {.(suc _)} {.(suc _)} (s≤s y≤z) | ⟨ x , y+x≡z ⟩ = ⟨ x , cong suc y+x≡z ⟩
```

I had to clean it up a bit.

5:25pm. Got distracted a little with the reply. Let me resume.

6:10pm. Done with lunch. That distracted me again.

```
∃-+-≤' : ∀ {y z} → y ≤ z → ∃[ x ] (y + x ≡ z)
∃-+-≤' {y} {z} z≤n = ⟨ z , refl ⟩
∃-+-≤' {y} {z} (s≤s y≤z) with ∃-+-≤' y≤z
∃-+-≤' {.(suc _)} {.(suc _)} (s≤s y≤z) | ⟨ x , y+x≡z ⟩ = ⟨ x , cong suc y+x≡z ⟩

∃-+-≤ : ∀ {y z} → y ≤ z → ∃[ x ] (x + y ≡ z)
∃-+-≤ {y} {z} y≤z with ∃-+-≤' y≤z
∃-+-≤ {y} {z} y≤z | ⟨ x , x₁ ⟩ rewrite +-comm y x = ⟨ x , x₁ ⟩
```

Literally figuring out how to do this rewrite was harder than the problem itself. I need to redo it so it is done properly.

No need for two proofs.

After that I will call it a day.

6:20pm. Agh, no it is still a pain.

6:25pm. No do not want to waste anymore time figuring out how to swap a bunch of arguments. Screw this.

This would be easy in Coq, but is a pain in the ass in Agda. It is not important.

6:30pm. There are two exercises left now. Unfortunately, I could not do the whole chapter. Admittedly the private Reddit posting I've been doing distracted me, but I never regret such interactions. They are rare as it is anyway.

6:35pm. I'll finish the chapter tomorrow properly and then move on to the next one.

Let me get on with the chilling.

6:35pm. https://stackoverflow.com/questions/56304634/is-functional-extensionality-with-dependent-functions-consistent/56305945

Got a reply to my earlier SO question. The thing compiles. I'll study it later.

https://stackoverflow.com/questions/39239363/what-is-axiom-k/39256457

I'll check out Axiom K at a later point as well."