"3:35pm.
```
#### Exercise `reverse-++-commute` (recommended)

Show that the reverse of one list appended to another is the
reverse of the second appended to the reverse of the first:
\begin{code}
postulate
  reverse-++-commute : ∀ {A : Set} {xs ys : List A}
    → reverse (xs ++ ys) ≡ reverse ys ++ reverse xs
\end{code}

#### Exercise `reverse-involutive` (recommended)

A function is an _involution_ if when applied twice it acts
as the identity function.  Show that reverse is an involution:
\begin{code}
postulate
  reverse-involutive : ∀ {A : Set} {xs : List A}
    → reverse (reverse xs) ≡ xs
\end{code}
```

Let me give these two a shot. They won't be hard. I did all this stuff in Coq.

I am already 1/3rd into the list chapter. I might not finish it today, but I will definitely finish it tomorrow if that turns out to be the case.

3:45pm.

```
reverse-++-commute : ∀ {A : Set} {xs ys : List A}
  → reverse (xs ++ ys) ≡ reverse ys ++ reverse xs
reverse-++-commute {A} {[]} {ys} rewrite ++-identityʳ (reverse ys) = refl
reverse-++-commute {A} {x ∷ xs} {ys} with reverse-++-commute {A} {xs} {ys}
... | prf = let prf' = cong (_++ [ x ]) prf in {!!}
```

```
Goal: reverse (xs ++ ys) ++ [ x ] ≡
      reverse ys ++ reverse xs ++ [ x ]
————————————————————————————————————————————————————————————
prf' : reverse (xs ++ ys) ++ [ x ] ≡
       (reverse ys ++ reverse xs) ++ [ x ]
x    : A
prf  : reverse (xs ++ ys) ≡ reverse ys ++ reverse xs
ys   : List A
xs   : List A
A    : Set
```

Ah, I see why this is giving me trouble. How annoying.

4pm.

```
reverse-++-commute : ∀ {A : Set} {xs ys : List A}
  → reverse (xs ++ ys) ≡ reverse ys ++ reverse xs
reverse-++-commute {A} {[]} {ys} rewrite ++-identityʳ (reverse ys) = refl
reverse-++-commute {A} {x ∷ xs} {ys} with reverse-++-commute {A} {xs} {ys}
... | prf with cong (_++ [ x ]) prf
... | prf' rewrite ++-assoc (reverse ys) (reverse xs) [ x ] = prf'
```

This one is done. Let me do the next one and I will take a break.

My focus is starting to waver at this point.

4:05pm.

```
reverse-involutive : ∀ {A : Set} {xs : List A}
  → reverse (reverse xs) ≡ xs
reverse-involutive {A} {[]} = refl
reverse-involutive {A} {x ∷ xs} rewrite reverse-++-commute {A} {reverse xs} {[ x ]}= cong (x ∷_) reverse-involutive
```

This was easy.

```
shunt-reverse : ∀ {A : Set} (xs ys : List A)
  → shunt xs ys ≡ reverse xs ++ ys
shunt-reverse [] ys =
  begin
    shunt [] ys
  ≡⟨⟩
    ys
  ≡⟨⟩
    reverse [] ++ ys
  ∎
shunt-reverse (x ∷ xs) ys =
  begin
    shunt (x ∷ xs) ys
  ≡⟨⟩
    shunt xs (x ∷ ys)
  ≡⟨ shunt-reverse xs (x ∷ ys) ⟩
    reverse xs ++ (x ∷ ys)
  ≡⟨⟩
    reverse xs ++ ([ x ] ++ ys)
  ≡⟨ sym (++-assoc (reverse xs) [ x ] ys) ⟩
    (reverse xs ++ [ x ]) ++ ys
  ≡⟨⟩
    reverse (x ∷ xs) ++ ys
  ∎
```

This problem gave me a decent bit of trouble in Coq. It was the first advanced 4 star problem that I did on one of the chapters.

Amazing how simple it came out in Agda.

4:10pm.

```
#### Exercise `map-compose`

Prove that the map of a composition is equal to the composition of two maps:
\begin{code}
postulate
  map-compose : ∀ {A B C : Set} {f : A → B} {g : B → C}
    → map (g ∘ f) ≡ map g ∘ map f
\end{code}
The last step of the proof requires extensionality.
```

Let me stop here for a bit. There is a decent number of exercises in this chapter, but none of them seem extremely difficult. I'll be able to handle it somehow."