I meant to start over 1.5 hours ago and do some work, but ended up talking to Anti. It is really a pity the correspondence is private as it allowed me to air some of my thoughts that I've had lying around.
I will keep the interaction in mind when it comes time to do the Simulacrum blog update and hopefully incorporate some of the material from that.

Let me take a break here. Some days are just like this. I did a lot yesterday and feel like lazing around.

I am annoyed that I did not get a reply to the local rewrite question I posed.

I'll just skip the exercises that need that for the time being. I have no intention of using Agda after all until that is done. I'll master Coq when I get back to it.

Today I will finally finish that book and get to studying the pigeonhole proof. I've been looking forward to that.

There is not much left of the book.