"4:50pm. 5/17. I just do not have the same feel as programming when I am reading a paper. I am going to get the code that comes along with the paper and play with it. Tomorrow. I really am quite tired at this point in the day.
5:15pm. 8/17. The `<->` operator on these pages confuses me. It can't possibly be that all of this is equivalent. What is going on here?

5:40pm. http://www.cse.chalmers.se/~nad/publications/danielsson-bag-equivalence.html

Got the Github repo accompanying the paper. It seems like there is a lot of stuff there.

https://homotopytypetheory.org/book/

I might as well get the PDF for this while I am at it.

5:45pm. I really have a lot of work ahead of me before I am done. At this rate I will be lucky to finish the Agda book before the month is up. I will leave this problem aside tomorrow and focus on the other ones for the time being. On the 31st I will take a break to do that monthly review.

All in all, I did quite a bit today. Noteworthy is doing the arithmetic series proof. This is the first 'real' math proof I've ever done in a dependently typed language.

```
#### Exercise `any?` (stretch)
#### Exercise `All-∀`
#### Exercise `Any-∃`
#### Exercise `filter?` (stretch)
```

Besides the two I've decided to skip for the time being, only these 4 are left. I will focus on them tomorrow. Then comes the bag paper. It is interesting that he proves correctness of parser combinators. Regexp parsers are what is giving me trouble in Coq as well, so I might be able lift something from this.

6pm. Ah, when I am done with this and the IndProp chapter, it will be such a relief. Just how much of my life is this going to take?

I need to get better. I need to get a lot better. Let me just skim the HoTT book for the first few pages and then I am done for the day."