"1:15pm. Let me resume. I could have started an hour ago, but I decided to rest a bit longer as I am so pissed off. I guess it is too much to expect a steady stream of positive reinforcement as I go into the harder problems.
From here on out there will be a change. Studying the code for that paper will be the first time I will be seriously studying other people's programming. It has been a while since I last did this.

Well, I suppose it was not that long if one counts Pharo.  ...True, I have been doing it in various guises, but studying libraries of proofs will be a first.

This is where all the skill lies. If I can find the power in me to replicate all the advanced proofs that are in the wild, I will definitely find the power to find what I seek.

In that respect my plan is remarkably simple - level up the math skill and crack machine learning. With that in hand, that will allow me to start the prelude to the Singularity. It will allow me to scale up starting from small tasks to the big ones.

In retrospect, the last 4.5 years have been a hell that does not seem like it will end anytime soon. Spiral did level me up so it was not a complete waste, but it was still cowardice. I knew that algorithms uber alles, but I still wanted to go low level stuff out the way because I did not have confidence I would be able to do this. I guess it is true - I have a habit of not doing the right thing until I am backed against the wall. Mathematics is the backbone of everything. It is obvious that if I want to make this world my own, even if I lack talent I'd better develop it.

In the worst case, I might end up doing meaningless theorem proving from now till the end of my life. In the worst case, my talent might not turn out to be enough.

To me math is the ultimate despair. It is the complete absence of purpose. That is the only way I've able to see it since my school days. I envied the geniuses who had the spark for this.

I am going to have to force it. Force everything to go the way I want.

I cannot hope of the mediocrities of the research community to pull off a miracle. I need to do it myself.

1:35pm. Well, rather than starting here, let me just do the chores first. After that I'll have the whole day clearly ahead of me."