"8:50am. Rather than wasting time, let me just start on the monthly review for the next PL thread. I do not think it should be too long.
"Going into May, I spent the last week of April going through the [Software Foundations](https://softwarefoundations.cis.upenn.edu/lf-current/toc.html). I had a very hard time with this book back in 2016 and finally three years later, I've been able to verify whether all that programming practice in the making of [Spiral](https://github.com/mrakgr/The-Spiral-Language/tree/v0.1) made any difference. I am definitely much better than I was 2.5 years ago, of that I am sure.

I've made pretty good progress for the first 6 chapters at a rate of about 1-2 days per chapter, until I hit the [inductive propositions](https://softwarefoundations.cis.upenn.edu/lf-current/IndProp.html) one. Since the start of June I've worked for nearly 3 weeks on it, and I still haven't finished it. It is just so insanely hard. The large quantities of very hard, mostly optional exercises that I keep getting stuck on really ground my progress to a snail's pace. I intend to resume it once I am done writing this review.

Just yesterday I've finished going through the [Agda version](https://plfa.github.io/) of the SF book. Baring the two exercises that I could not do due to compiler issues, I can claim 100% completion. Unlike the Coq version, this one does not have random high difficulty exercises sprinkled throughout so it was much easier. Admittedly, a language like Agda is much closer to my preferred programming style than Coq styled tactics, so that might play a part in it.

I agree with [Wadler's assesment](http://homepages.inf.ed.ac.uk/wadler/papers/sbmf/sbmf.pdf) of the differences between the two languages. Indeed the Agda proofs are much more elegant and readable than Coq ones. Elegance provides them the very important learning benefit of making them easier to commit to memory. And indeed in Agda you do not have to deal with Coq's induction nonsense.

That having said, at this point in time Agda has a definite lack of polish. The type errors are basically unreadable (surprisingly not a problem in Coq), the auto proof search will in a lot of cases output uncompilable code, the editor will sometimes not clear the buffer properly and display stale information in goals. It is also missing some automation opportunities. `let` statements do not allow pattern matching, case splitting is annoying in requiring me to {} wrap the argument in a lambda manually. In top level functions in Idris it is possible to make a template based on a type signature - Agda could learn from that rather than forcing me to copy every time. Compared to Coq, Agda definitely feels more sluggish while work in is progress in it.

For those reasons, I'd favor Coq at the moment if I had to use a theorem prover seriously.

---

May was such an uphill battle so far and that will continue for the foreseeable future.

The deep learning experiments of 2018 were the catalyst for this. Rather than being the solution to any question of intelligence, to me deep learning is in the end just a box full of mysteries. I've finally accepted that my old methods are not enough to solve anything and that gave me the desire and the wish to be good at math. I was always bad at this, but I guess given what I truly want to do, it was too much to hope that I would never be forced to come face to face with that adversary. Math is the backbone of modern civilization. If it was not this particular path, then it would have been something else that would make me undertake this quest. It was just too much to expect that I would be able to attain true understanding with a low skill in it. At its core, mathematics is the crystallization of understanding.

It sure does not seem like that when you are stuck on some trivial problem in Coq though.

I suppose just ignoring this and going with counter factual regret methods is an option when it comes to poker, but I know that won't get me anywhere to understanding computational learning. And if the point was money I'd be better of just getting a job instead.

Since things are like this and I have a lot of time, I'll just continue doing what I feel is right. I will master a formal system and merge the two disparate skills - math and programming into one. If I could get my math skill up to par with my programming, then that will give me a real chance of resolving the issues with backpropagation and making my own learning algorithm from first principles. The goal is to sharpen my intuition until I fill in all the gaps and find them.

Long term, it makes no sense for programming and mathematics to be two disparate skills and activities, but you can hardly blame me for thinking that way given that I started with Logo and Basic back in elementary school. Back then I was not yet a speculator."

10am. This is good enough. If I am lucky it will even fit into a single post. I've been overdoing it in the PL threads of the past few months with the size of them. One post is what I should strive for.

10:05am. I guess it is finally time I take a look at the pigeonhole proof by gallais.

No point in resting yet."